[seed_update] Enhance the generate and submit test functionality
In this diff we enhance the submit and generate test functionalities.

For the Generate test functionality we limit the number of questions to 10.
If a topic has less than 10 questions we return only however many questions
there are.

For the Submit functionalities we add feedback, as well as correct answer
if the answer given was not the correct one.

We also update the seed with more data.